requirements.txt
kaggle
jupyter
tqdm
ipython
numpy==1.23.5
matplotlib
graphviz
pydot
pandas
seaborn
scikit-learn==1.2.2
scikit-image==0.20.0
imagehash==4.3.1
tensorflow==2.12.0
tensorflow-addons==0.20.0
keras==2.12.0
keras-tuner==1.3.5
icrawler
pillow==9.5.0
joblib
huggingface
transformers
datasets
